refactor(barn, seeds): move addSeeds method to Seeds service, remove now unused Barn service
