+++ src/app/barn/seeds-api.service.ts
@@ -8,15 +8,28 @@ import type { DbSeed } from './rxdb/schema/seeds';
 export class SeedsApiService {
   private barnDbService = inject(BarnDbService);
 
-  getSeeds({ limit }: { limit?: number } = {}) {
+  getSeeds({ limit, names }: { limit?: number; names?: string[] } = {}) {
     return from(this.barnDbService.getDb()).pipe(
-      switchMap(db => db.seeds.find({ sort: [{ lastAddedAt: 'desc' }], limit }).$),
+      switchMap(
+        db =>
+          db.seeds.find({
+            ...(names ? { selector: { name: { $in: names } } } : {}),
+            sort: [{ lastAddedAt: 'desc' }],
+            limit,
+          }).$,
+      ),
       map(seeds => seeds.map(seed => seed.toMutableJSON())),
     );
   }
 
   getSeedsCount() {
     return from(this.barnDbService.getDb()).pipe(switchMap(db => db.seeds.count().$));
+  }
+
+  async addSeeds(seeds: DbSeed[]) {
+    const db = await this.barnDbService.getDb();
+
+    await db.seeds.bulkUpsert(seeds);
   }
 
   async updateSeed(name: string, newData: Partial<DbSeed>) {
